Add enableiOSAudiosession method
